sbotcli
port of
https://github.com/cryptix/cmd/commit/be616fe4f0f98375ed0bfd46bf04bffb3164f969
gophbot to the new muxrpc api

not a full port yet (see the todos) but its an example of how to dial
and create a client to call a remote.

+++ gossip/fetch.go
@@ -40,7 +40,7 @@ func (g *Handler) fetchFeed(ctx context.Context, fr sbot.FeedRef, e muxrpc.Endpo
 		Keys:  false,
 		Live:  false,
 		Id:    fr.Ref(),
-		Seq:   latestSeq + 1,
+		Seq:   int64(latestSeq + 1),
 		Limit: 1000,
 	}
 	start := time.Now()
